Assignment 2 (2v)
In this version 2, I had made an return button in the quiz question for users to click on if they want to return to main menu screen. Once press, users will see if another text box confirming if the users want to return or not. The button did got shown, but currently it won't switch back to the main menu screen. For the next version, we will fix the return menu button and work on other features.